+++ lib/src/core/constants.dart
@@ -1,0 +1,8 @@
+/// Safe padding for context menu to not be too close to the screen edges.
+const double kContextMenuSafePadding = 8.0;
+
+/// The default menu item height.
+const double kMenuItemHeight = 32.0;
+
+/// The default menu item icon size.
+const double kMenuItemIconSize = 32.0;